feat(breadcrumb): try to add behavior to other templates
