+++ src/generic/generic-component-breadcrumb/package.json
@@ -8,9 +8,9 @@
   "license": "EUPL-1.1",
   "dependencies": {
     "@ecl/generic-base": "^1.0.0",
-    "@ecl/generic-component-expandable": "^1.0.0",
     "@ecl/generic-component-link": "^1.0.0",
-    "@ecl/generic-utility-screen-reader": "^1.0.0"
+    "@ecl/generic-utility-screen-reader": "^1.0.0",
+    "lodash.debounce": "4.0.8"
   },
   "devDependencies": {
     "@ecl/generic-utility-aria": "^1.0.0"
