+++ src/systems/ec/ec-component/ec-component-breadcrumb/package.json
@@ -8,12 +8,9 @@
   "license": "EUPL-1.1",
   "dependencies": {
     "@ecl/ec-base": "^1.0.0",
-    "@ecl/ec-component-expandable": "^1.0.1",
     "@ecl/ec-component-link": "^1.0.1",
     "@ecl/ec-utility-screen-reader": "^1.0.0",
-    "@ecl/generic-base": "^1.0.0",
-    "@ecl/generic-component-breadcrumb": "^1.0.0",
-    "lodash.debounce": "4.0.8"
+    "@ecl/generic-component-breadcrumb": "^1.0.0"
   },
   "devDependencies": {
     "@ecl/ec-utility-aria": "^1.0.0"
