+++ src/systems/ec/ec-component/ec-component-page-header/ec-component-page-header.config.js
@@ -29,4 +29,12 @@ module.exports = {
       },
     },
   ],
+  context: {
+    _demo: {
+      scripts: `
+        document.addEventListener('DOMContentLoaded', function () { 
+          ECL.initBreadcrumb();
+        });`,
+    },
+  },
 };
