+++ src/systems/eu/eu-component/eu-component-breadcrumb/package.json
@@ -12,6 +12,9 @@
     "@ecl/eu-utility-screen-reader": "^1.0.0",
     "@ecl/generic-component-breadcrumb": "^1.0.0"
   },
+  "devDependencies": {
+    "@ecl/ec-utility-aria": "^1.0.0"
+  },
   "publishConfig": {
     "access": "public"
   },
